Move logic to the separate TodoItem component
